Initial implementation (#1)
This enables calling the original Sleigh compiler implementation in the
Ghidra repository as a library. This required some minor changes to the
original Sleigh source code since it expected to be invoked as an
executable. These changes are documented in the forked `slgh_compile.cc`
file.

This initial commit is based on Ghidra release 10.4.